🌱 split acmm: extract data into separate module
